requirements.txt
robotframework == 6.0.2
robotframework-seleniumlibrary == 6.0.0
beautifulsoup4 == 4.11.1
robotframework-screencaplibrary == 1.6.0